Update readme.md
It is no longer a goal of the project to have ; on the same line as object names for the purpose of indicating the end of the object because after thinking about it, it's not worth the time and it wouldn't be really useful. Personally, I don't even see the point of empty objects, therefore, I've also removed that pointless example.